+++ resale/src/main/java/com/goti/resale/repository/order/ResaleOrderRepository.java
@@ -1,4 +1,4 @@
-package com.goti.resale.repository;
+package com.goti.resale.repository.order;
 
 import java.util.UUID;
