+++ resale/src/main/java/com/goti/resale/repository/order/ResaleOrderRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.goti.resale.repository;
+package com.goti.resale.repository.order;
 
 import java.time.LocalDate;
 import java.util.List;
