+++ resale/src/main/java/com/goti/resale/repository/trasaction/ResaleTransactionRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.goti.resale.repository;
+package com.goti.resale.repository.trasaction;
 
 import java.util.List;
 import java.util.UUID;
